refactor tiff
