+++ modules/geotiff/src/index.ts
@@ -1,3 +1,3 @@
-export { loadOmeTiff } from './tiff';
-
-export { default as TiffPixelSource } from './tiff/pixel-source';
+export {loadGeoTiff} from './lib/load-geotiff';
+export {loadOMETiff} from './lib/load-ome-tiff';
+export {default as TiffPixelSource} from './lib/tiff-pixel-source';
