+++ modules/geotiff/src/lib/load-geotiff.ts
@@ -5,10 +5,10 @@ import {
   // createPoolProxy,
   createOffsetsProxy,
   checkProxies
-} from './lib/proxies';
+} from './utils/proxies';
 // import Pool from './lib/Pool';
 
-import { load } from './ome-tiff';
+import { loadOMETiff } from './load-ome-tiff';
 
 interface TiffOptions {
   headers?: Record<string, unknown>;
@@ -26,7 +26,7 @@ interface TiffOptions {
  * multi-threaded pool of image decoders should be used to decode tiles (default = true).
  * @return {Promise<{ data: TiffPixelSource[], metadata: ImageMeta }>} data source and associated OME-Zarr metadata.
  */
-export async function loadOmeTiff(source: string | File, opts: TiffOptions = {}) {
+export async function loadGeoTiff(source: string | File, opts: TiffOptions = {}) {
   const { headers, offsets, pool = true } = opts;
 
   let tiff: GeoTIFF;
@@ -61,5 +61,5 @@ export async function loadOmeTiff(source: string | File, opts: TiffOptions = {})
    */
   checkProxies(tiff);
 
-  return load(tiff);
+  return loadOMETiff(tiff);
 }
