+++ modules/geotiff/src/lib/ome/omexml.ts
@@ -1,5 +1,5 @@
 import parser from 'fast-xml-parser';
-import { ensureArray, intToRgba } from './utils';
+import { ensureArray, intToRgba } from '../utils/tiff-utils';
 
 // WARNING: Changes to the parser options _will_ effect the types in types/omexml.d.ts.
 const PARSER_OPTIONS = {
