+++ modules/geotiff/src/lib/utils/Pool.ts
