+++ modules/geotiff/src/lib/utils/decoder.worker.ts.disabled
