+++ modules/geotiff/src/lib/utils/proxies.ts
