+++ modules/geotiff/test/index.js
@@ -1,4 +1,3 @@
 import './utils.spec';
-
-import './ome-tiff.spec';
+import './tiff.spec';
 // import './tiff-lib.spec';
